[ELG-279] resolve cells vs vaps field ID conflict
The master branch added the new cells field with ID 10, but develop
branch had previously added the vaps field with ID 10. The master branch
(with the new cells field) has already been deployed to production,
while the develop branch (with the new vaps field) has not. So - the
cells field wins, and its ID remains 10. The vaps ID is now 11.