Change type of sku field from bytes to string
This change should be transparant to the client, since protobuf strings
are encoded in the same way as bytes. But it makes a difference to the
server: a protobuf bytes field would need to be explicitly converted to
a string (e.g., when valuing the Skyhook-SKU header field), whereas a
protobuf string field is just, well, a string, and requires no such
conversion.